refactor: restructure Settings module
Settings now has two subpackages: api & impl. Code is currently in impl.
