+++ feature/settings/impl/src/main/AndroidManifest.xml
