+++ feature/settings/impl/src/main/res/mipmap-mdpi/ic_launcher.webp
