+++ feature/settings/impl/src/main/res/mipmap-xxhdpi/ic_launcher.webp
