+++ settings.gradle.kts
@@ -48,7 +48,8 @@ include(":feature:search:api")
 include(":core:markdown")
 include(":feature:video")
 include(":core:domain")
-include(":feature:settings")
+include(":feature:settings:impl")
+include(":feature:settings:api")
 include(":core:datastore")
 include(":core:datastore-proto")
 include(":core:database")
